+++ projects/test_models/stereotypes/src/test_data/multi_concept_model_td.cpp
@@ -24,7 +24,7 @@
 namespace {
 
 int create_int(const unsigned int position) {
-    return position;
+    return static_cast<int>(position);
 }
 
 std::string create_std_string(const unsigned int position) {
